refactor(blob-export): move gate constants to blob-export-gate.ts
index.ts was a display-constants file (dropdown labels, field group
descriptions). LEGACY_BLOB_EXPORT_CUTOFF, LEGACY_BLOB_EXPORT_SOURCES,
and isLegacyBlobExportAllowed are business-rule constants — a different
concern. Extract them to a dedicated blob-export-gate.ts co-located in
the same directory; index.ts re-exports via `export *` so all import
paths remain unchanged.
